Aggressively simplify device::Device.
- Conversions of Device from/to libc dev_t and conversion of Device
  from a string were removed.  None of these were actually in use.
  (Some may come back later.)

- Device::from_kdev_t now takes a u64 and interprets it according to the
  C library’s extended 64-bit dev_t format.  This format is consistent
  among all three of GNU libc, musl libc, and Bionic libc, and agrees
  with the kernel’s 32-bit format for all values that fit in 32 bits,
  so it’s the most probable way for the Linux kernel’s kdev_t to be
  extended to 64 bits in the future.  This eliminates a few more casts
  in dm.rs.

- Device::to_kdev_t still returns None if asked to encode a Device
  with numbers that don’t fit in 32-bit kdev_t.

+++ src/deviceinfo.rs
@@ -67,9 +67,7 @@ impl TryFrom<dmi::Struct_dm_ioctl> for DeviceInfo {
             open_count: ioctl.open_count,
             flags: DmFlags::from_bits_truncate(ioctl.flags),
             event_nr: ioctl.event_nr,
-            // dm_ioctl struct reserves 64 bits for device but kernel "huge"
-            // encoding is only 32 bits.
-            dev: Device::from_kdev_t(ioctl.dev as u32),
+            dev: Device::from_kdev_t(ioctl.dev),
             uuid,
             name,
         })
